fix: 6 add syncdataset to get metrics (#7)
* feat: add sync_dataset module for sync reads.

* feat: use sync_dataset for sync file handling.

* feat: modify utils module for sync_dataset handling.

* Edited sync_dataset.py

* fix: remove sync_dataset.

* fix: revert calls to sync_dataset

* fix: revert calls to sync_dataset and cast line bits to int64 type

+++ code/utils.py
@@ -263,7 +263,7 @@ def get_bit_changes(sync_file, bit):
         Bit for which to return changes.
 
     """
-    bit_array = get_sync_file_bit(sync_file, bit)
+    bit_array = get_sync_file_bit(sync_file, bit).astype(np.int64)
     return np.ediff1d(bit_array, to_begin=0)
